test(cli): consolidate inline-imports tests into one file (#1141)
The split between `inline-imports-collision.test.ts` and
`inline-imports-ast.test.ts` was an authoring artefact — both files
exercise the same `resolveRelativeImports` function and the same IIFE-wrap
behaviour. Splitting on "core" vs "AST edge cases" suggests they cover
different code paths, which they don't. Merge into one descriptive file
and update the leading comment to reflect the full scope.

13 tests, no behavior change.
