+++ packages/cli/src/__tests__/inline-imports-collision.test.ts
@@ -1,11 +1,13 @@
-// Regression tests for piconic-ai/barefootjs#1141 — when two sibling .ts
-// modules each declare a module-private const/function with the same name,
-// the inliner used to splice both decls at top level, producing
-// `SyntaxError: Identifier '...' has already been declared` in the browser.
+// Tests for piconic-ai/barefootjs#1141 — `resolveRelativeImports` wraps
+// each inlined module's body in an IIFE that re-exports only the names
+// the parent imports, so two siblings each declaring a module-private
+// const with the same identifier (e.g. `BAR_STYLE` in both `CanvasActions.ts`
+// and `ZoomBar.ts`) no longer collide at the parent's top level.
 //
-// The fix wraps each inlined module's body in an IIFE that re-exports only
-// the names the parent imports, so module-private decls stay scoped to
-// their own arrow body.
+// Covers the core IIFE-wrap behaviour AND the edge cases the
+// TypeScript-AST-based parsing handles (multi-line / trailing-comma /
+// commented imports & exports, type-only decls, string literals containing
+// `import` / `export`).
 
 import { describe, test, expect, beforeEach, afterAll } from 'bun:test'
 import { resolveRelativeImports } from '../lib/resolve-imports'
@@ -31,7 +33,7 @@ function countMatches(s: string, re: RegExp): number {
   return [...s.matchAll(re)].length
 }
 
-describe('resolveRelativeImports — module-private name collisions (bf#1141)', () => {
+describe('resolveRelativeImports — IIFE-wrapped inlining (bf#1141)', () => {
   test('two siblings sharing a private const: at most one TOP-LEVEL decl', async () => {
     writeFileSync(
       resolve(COMPONENTS_DIR, 'CanvasActions.ts'),
@@ -208,5 +210,229 @@ console.log(ns.a, ns.b(), new ns.C())
     expect(result).toMatch(/return \{[^}]*\bC\b[^}]*\}/)
     // HIDDEN (private) is not in the return.
     expect(result).not.toMatch(/return \{[^}]*\bHIDDEN\b/)
+  })
+
+  // ── AST edge cases ────────────────────────────────────────────────────────
+  // Shapes the TypeScript-AST parser handles correctly that a regex pass
+  // would miss or misclassify.
+
+  test('multi-line `export { … }` block is fully stripped', async () => {
+    writeFileSync(
+      resolve(COMPONENTS_DIR, 'multi-export.ts'),
+      `function alpha() { return 1 }
+function beta() { return 2 }
+function gamma() { return 3 }
+export {
+  alpha,
+  beta,
+  gamma,
+}
+`,
+    )
+    const clientJs = `import * as ns from './multi-export'
+console.log(ns.alpha(), ns.beta(), ns.gamma())
+`
+    writeFileSync(resolve(COMPONENTS_DIR, 'Comp-me.js'), clientJs)
+
+    const manifest = {
+      Comp: { clientJs: 'components/canvas/Comp-me.js', markedTemplate: 'components/canvas/Comp.tsx' },
+    }
+
+    await resolveRelativeImports({ distDir: DIST_DIR, manifest })
+
+    const result = await Bun.file(resolve(COMPONENTS_DIR, 'Comp-me.js')).text()
+    // The IIFE return surfaces all three (collected from the original TS
+    // source's `export {…}` block).
+    expect(result).toMatch(/return \{[^}]*\balpha\b[^}]*\}/)
+    expect(result).toMatch(/return \{[^}]*\bbeta\b[^}]*\}/)
+    expect(result).toMatch(/return \{[^}]*\bgamma\b[^}]*\}/)
+    // The multi-line block must not survive in the body.
+    expect(result).not.toMatch(/^\s*export\s*\{/m)
+    expect(() => new Function(result)).not.toThrow()
+  })
+
+  test('trailing comma in `export { a, }` is handled', async () => {
+    writeFileSync(
+      resolve(COMPONENTS_DIR, 'trail.ts'),
+      `function onlyOne() { return 42 }
+export {
+  onlyOne,
+}
+`,
+    )
+    const clientJs = `import * as ns from './trail'
+console.log(ns.onlyOne())
+`
+    writeFileSync(resolve(COMPONENTS_DIR, 'Comp-tc.js'), clientJs)
+
+    const manifest = {
+      Comp: { clientJs: 'components/canvas/Comp-tc.js', markedTemplate: 'components/canvas/Comp.tsx' },
+    }
+
+    await resolveRelativeImports({ distDir: DIST_DIR, manifest })
+
+    const result = await Bun.file(resolve(COMPONENTS_DIR, 'Comp-tc.js')).text()
+    expect(result).toMatch(/return \{\s*onlyOne\s*\}/)
+    expect(result).not.toMatch(/^\s*export\s*\{/m)
+    expect(() => new Function(result)).not.toThrow()
+  })
+
+  test('aliased export `export { a as b }` collects the exported alias', async () => {
+    writeFileSync(
+      resolve(COMPONENTS_DIR, 'alias-exp.ts'),
+      `function internalName() { return 'aliased' }
+export { internalName as publicName }
+`,
+    )
+    const clientJs = `import * as ns from './alias-exp'
+console.log(ns.publicName())
+`
+    writeFileSync(resolve(COMPONENTS_DIR, 'Comp-ax.js'), clientJs)
+
+    const manifest = {
+      Comp: { clientJs: 'components/canvas/Comp-ax.js', markedTemplate: 'components/canvas/Comp.tsx' },
+    }
+
+    await resolveRelativeImports({ distDir: DIST_DIR, manifest })
+
+    const result = await Bun.file(resolve(COMPONENTS_DIR, 'Comp-ax.js')).text()
+    expect(result).toMatch(/return \{\s*publicName\s*\}/)
+    expect(result).not.toMatch(/return \{\s*internalName\b/)
+  })
+
+  test('type-only export is excluded from namespace IIFE return', async () => {
+    writeFileSync(
+      resolve(COMPONENTS_DIR, 'types-mixed.ts'),
+      `export interface Shape { x: number }
+export type Color = 'red' | 'blue'
+export const valueExport = 1
+export function fnExport() { return 2 }
+type Internal = string
+export type { Internal }
+`,
+    )
+    const clientJs = `import * as ns from './types-mixed'
+console.log(ns.valueExport, ns.fnExport())
+`
+    writeFileSync(resolve(COMPONENTS_DIR, 'Comp-tm.js'), clientJs)
+
+    const manifest = {
+      Comp: { clientJs: 'components/canvas/Comp-tm.js', markedTemplate: 'components/canvas/Comp.tsx' },
+    }
+
+    await resolveRelativeImports({ distDir: DIST_DIR, manifest })
+
+    const result = await Bun.file(resolve(COMPONENTS_DIR, 'Comp-tm.js')).text()
+    expect(result).toMatch(/return \{[^}]*\bvalueExport\b[^}]*\}/)
+    expect(result).toMatch(/return \{[^}]*\bfnExport\b[^}]*\}/)
+    // Type-only names have no runtime binding — must not appear in the return.
+    expect(result).not.toMatch(/return \{[^}]*\bShape\b[^}]*\}/)
+    expect(result).not.toMatch(/return \{[^}]*\bColor\b[^}]*\}/)
+    expect(result).not.toMatch(/return \{[^}]*\bInternal\b[^}]*\}/)
+    expect(() => new Function(result)).not.toThrow()
+  })
+
+  test('string literal containing `import` / `export` is not stripped', async () => {
+    writeFileSync(
+      resolve(COMPONENTS_DIR, 'str-lit.ts'),
+      "const docs = 'how to import this module: see README'\n" +
+      "const more = 'export this if needed'\n" +
+      "export function getDocs() { return docs + ' / ' + more }\n",
+    )
+    const clientJs = `import { getDocs } from './str-lit'
+console.log(getDocs())
+`
+    writeFileSync(resolve(COMPONENTS_DIR, 'Comp-sl.js'), clientJs)
+
+    const manifest = {
+      Comp: { clientJs: 'components/canvas/Comp-sl.js', markedTemplate: 'components/canvas/Comp.tsx' },
+    }
+
+    await resolveRelativeImports({ distDir: DIST_DIR, manifest })
+
+    const result = await Bun.file(resolve(COMPONENTS_DIR, 'Comp-sl.js')).text()
+    // String literals must survive intact — they're not import/export
+    // statements, just strings.
+    expect(result).toContain('how to import this module')
+    expect(result).toContain('export this if needed')
+    expect(() => new Function(result)).not.toThrow()
+  })
+
+  test('comment inside named export decl: `export /* keep */ const a = 1`', async () => {
+    writeFileSync(
+      resolve(COMPONENTS_DIR, 'cmt-exp.ts'),
+      `export /* keep */ const valueA = 'a'
+export const valueB = 'b'
+`,
+    )
+    const clientJs = `import { valueA, valueB } from './cmt-exp'
+console.log(valueA, valueB)
+`
+    writeFileSync(resolve(COMPONENTS_DIR, 'Comp-cx.js'), clientJs)
+
+    const manifest = {
+      Comp: { clientJs: 'components/canvas/Comp-cx.js', markedTemplate: 'components/canvas/Comp.tsx' },
+    }
+
+    await resolveRelativeImports({ distDir: DIST_DIR, manifest })
+
+    const result = await Bun.file(resolve(COMPONENTS_DIR, 'Comp-cx.js')).text()
+    expect(result).toMatch(/const \{\s*valueA,\s*valueB\s*\}\s*=\s*\(\(\) =>/)
+    expect(result).not.toMatch(/^\s*export\s/m)
+    expect(() => new Function(result)).not.toThrow()
+  })
+
+  test('type-only import in inlined module is erased by transpile', async () => {
+    writeFileSync(
+      resolve(COMPONENTS_DIR, 'types-only.ts'),
+      `export type Color = 'red' | 'blue'
+`,
+    )
+    writeFileSync(
+      resolve(COMPONENTS_DIR, 'consumer.ts'),
+      `import type { Color } from './types-only'
+export const pickColor = (c: Color) => c
+`,
+    )
+    const clientJs = `import { pickColor } from './consumer'
+console.log(pickColor('red'))
+`
+    writeFileSync(resolve(COMPONENTS_DIR, 'Comp-to.js'), clientJs)
+
+    const manifest = {
+      Comp: { clientJs: 'components/canvas/Comp-to.js', markedTemplate: 'components/canvas/Comp.tsx' },
+    }
+
+    await resolveRelativeImports({ distDir: DIST_DIR, manifest })
+
+    const result = await Bun.file(resolve(COMPONENTS_DIR, 'Comp-to.js')).text()
+    expect(result).toMatch(/const \{\s*pickColor\s*\}\s*=\s*\(\(\) =>/)
+    expect(result).not.toMatch(/^\s*import\s/m)
+    expect(result).not.toMatch(/^\s*export\s/m)
+    expect(() => new Function(result)).not.toThrow()
+  })
+
+  test('named import with non-trivial whitespace is parsed', async () => {
+    writeFileSync(
+      resolve(COMPONENTS_DIR, 'wide.ts'),
+      `export const aaa = 1
+export const bbb = 2
+`,
+    )
+    const clientJs = `import {   aaa as renamedA  ,   bbb   } from './wide'
+console.log(renamedA, bbb)
+`
+    writeFileSync(resolve(COMPONENTS_DIR, 'Comp-wd.js'), clientJs)
+
+    const manifest = {
+      Comp: { clientJs: 'components/canvas/Comp-wd.js', markedTemplate: 'components/canvas/Comp.tsx' },
+    }
+
+    await resolveRelativeImports({ distDir: DIST_DIR, manifest })
+
+    const result = await Bun.file(resolve(COMPONENTS_DIR, 'Comp-wd.js')).text()
+    expect(result).toMatch(/const \{\s*aaa:\s*renamedA,\s*bbb\s*\}\s*=\s*\(\(\) =>/)
+    expect(result).toMatch(/return \{\s*aaa,\s*bbb\s*\}/)
+    expect(() => new Function(result)).not.toThrow()
   })
 })
